[WIP] fix: change Response.insertion == nil to act like []
I don't remember the full context around the previous PR.  promoted-ts-client treats nil and [] as the same.  I didn't see any Delivery API logic for rendering out insertion == nil as an empty array.  Protos should default not set and empty array to the same value.  If Delivery API returns a successful deliver call and insertions is nil, I think that should be treated like an empty array.  Delivery API needs to make sure to send an error if it actually hit an error so SDKs can handle it.

It's okay for the offset>=len(items) case to exist with this because, if we assume it's valid input, it'd return an empty list of response insertions anyways.

The current behavior (if response.insertion==nil, then return fullInsertions) is not good and will likely break eventually.  For deliver, we want to return a paged list of items.

This is a partial rollback to the previous way but converts nil as [].

TESTING=rspec